move expression to core
